coroutines: Use cp_build_init_expr consistently.
Tested on x86_64-darwin19, OK for master?
thanks
Iain

-- >8 --

Now we have the TARGET_EXPR_ELIDING_P flag, it is important to ensure it
is set properly on target exprs.  The code here has a mixture of APIs used
to build inits.  This patch changes that to use cp_build_init_expr() where
possible, since that handles setting the flag.