ifcvt: Fix bitpos calculation in bitfield lowering [PR107229]
Hi,

The bitposition calculation for the bitfield lowering in loop if
conversion was not
taking DECL_FIELD_OFFSET into account, which meant that it would result in
wrong bitpositions for bitfields that did not end up having representations
starting at the beginning of the struct.

Bootstrappend and regression tested on aarch64-none-linux-gnu and
x86_64-pc-linux-gnu.

gcc/ChangeLog:

     PR tree-optimization/107229
     * gcc/tree-if-conv.cc (get_bitfield_rep): Fix bitposition calculation.

gcc/testsuite/ChangeLog:

     * gcc.dg/vect/pr107229-1.c: New test.
     * gcc.dg/vect/pr107229-2.c: New test.
     * gcc.dg/vect/pr107229-3.c: New test.