libgomp: fix hang on fatal error
This patch fixes a problem in which fatal errors inside mutex-locked
regions (i.e. basically anything in the plugin) will cause it to hang up
trying to take the lock to clean everything up.

Using abort() instead of exit(1) bypasses the atexit handlers and solves
the problem.

OK for mainline?

Andrew
libgomp: fix hang on fatal error

Don't try to clean up if a fatal error occurs in libgomp.  Typically the
cleanup is not reentrant so we end up hung on a lock.

libgomp/ChangeLog:

	* error.c (gomp_vfatal): Use abort instead of exit.