RISC-V: Implement __clear_cache via __builtin___clear_cache
We have had an implementation of __builtin___clear_cache since the
beginning, but didn't have the cooresponding __clear_cache library
routine implemented.  This directly conflicts the GCC manual in a
handful of places, which indicates that __clear_cache should work and
that __builtin_clear_cache should function the same way as __clear_cache
by ethier calling it or inlining the functionality.

This patch simply implements __clear_cache via __builtin___clear_cache.
This should be safe as we always have clear_cache insn so therefor
__builtin___clear_cache will never fall back to calling __clear_cache.
I'm not actually sure that silently implementing clear_cache as a NOP
when there is no ISA defined mechanism for icache synchronization is the
right way to go, but that's really a different discussion.

This was reported as Bug 94136, which is a year old but was categorized
as a documentation bug.  I believe that categorization was incorrect:
having an empty __clear_cache library routine is simply incorrect
behavior, the fact that __builtin___clear_cache happens to be
implemented as a libc call on Linux is just a red herring suggesting the
documentation fix to point out the name difference.  I view this new
behavior as conforming to the existing documentation: we're just
inlining the __clear_cache implementation, even if that implementation
happens to be a call to a very similar looking libc routine.

gcc/ChangeLog
	PR target/94136
	* config/riscv/riscv.h (CLEAR_INSN_CACHE): New macro.