Don't print discriminators for -fcompare-debug.
With -gstatement-frontiers we may end up with different IR
coming from the front end with and without debug information turned on.
See https://gcc.gnu.org/bugzilla/show_bug.cgi?id=100733 for details.
That may result in differences in discriminator values and -fcompare-debug
failures.

This patch disables printing of discriminators when the dump is intended
for -fcompare-debug comparison and reverses the workaround in a test.

Tested on x86_64-pc-linux-gnu.

gcc/ChangeLog:
	PR debug/107231
	PR debug/107169
	* print-rtl.cc (print_rtx_operand_code_i): Don't print discriminators
	for -fdebug-compare.

gcc/testsuite/ChangeLog:

	* c-c++-common/ubsan/pr85213.c: Reverse the workaround for discriminators.