Add missing splitter for H8
While testing a minor optimization on the H8 my builds failed due to
failure to split a zero-extended memory load.    That particular pattern
is a bit special on the H8 in that it's split at assembly time primarily
to get the length computations correct.  Arguably that alternative
should go away completely, but I haven't really looked into that.

Anyway, with the final-asm split we obviously need to match a
define_split somewhere.  But none was ever written after adding CCZN
optimizations.  So if we had a zero extend of a memory operand and it
was used to eliminate a compare, then we'd abort at final asm time.

Regression tested (in conjunction with various other in-progress
patches) on H8 without regressions.

Installed on the trunk.

Jeff
commit 43ee3f64cb519f2675fa1771007d4aa3baba944f
Author: Jeff Law <jeffreyalaw@gmail.com>
Date:   Mon Oct 17 19:19:25 2022 -0400

    Add missing splitter for H8

    While testing a minor optimization on the H8 my builds failed due to
    failure to split a zero-extended memory load.    That particular pattern
    is a bit special on the H8 in that it's split at assembly time primarily
    to get the length computations correct.  Arguably that alternative should
    go away completely, but I haven't really looked into that.

    Anyway, with the final-asm split we obviously need to match a define_split
    somewhere.  But none was ever written after adding CCZN optimizations.  So
    if we had a zero extend of a memory operand and it was used to eliminate
    a compare, then we'd abort at final asm time.

    Regression tested (in conjunction with various other in-progress patches) on
    H8 without regressions.

    gcc/
            * config/h8300/extensions.md (CCZN setting zero extended load): Add
            missing splitter.