SRA: Limit replacement creation for accesses propagated from LHSs
Hi,

PR 107206 is fallout from the fix to PR 92706 where we started
propagating accesses across assignments also from LHS to RHS of
assignments so that we would not do harmful total scalarization of the
aggregates on the RHS.

But this can lead to new scalarization of these aggregates and in the
testcase of PR 107206 these can appear in superfluous uses of
un-initialized values and spurious warnings.

Fixed by making sure the the accesses created by propagation in this
direction are only used as a basis for replacements when the structure
would be totally scalarized anyway.

Richi has already approved the patch (without the testcase) in Bugzilla,
so I plan to commit it after bootstrapping and testing it on an x86_64
(on top of a more recent trunk than last week).

Thanks,

Martin

gcc/ChangeLog:

2022-10-18  Martin Jambor  <mjambor@suse.cz>

	PR tree-optimization/107206
	* tree-sra.cc (struct access): New field grp_result_of_prop_from_lhs.
	(analyze_access_subtree): Do not create replacements for accesses with
	this flag when not toally scalarizing.
	(propagate_subaccesses_from_lhs): Set the new flag.

gcc/testsuite/ChangeLog:

2022-10-18  Martin Jambor  <mjambor@suse.cz>

	PR tree-optimization/107206
	* g++.dg/tree-ssa/pr107206.C: New test.