c++: Mitigate -Wuseless-cast with classes [PR85043]
-Wuseless-cast (not part of -Wall/-Wextra) warns here:

  struct S { };
  void g (S&&);
  void f (S&& arg)
  {
    g (S(arg)); // warning: useless cast to type 'struct S'
  }

which is wrong: the code will not compile without the cast because
"arg" is an lvalue which cannot bind to S&&.

I'd like to disable the warning when a class object is cast to
a non-reference type, which seems like a minimal patch that fixes the
problems reported in our Bugzilla.  Of course, the cast in "(int)i"
may not be useless, either, but I'm not changing that here.

Bootstrapped/regtested on x86_64-pc-linux-gnu, ok for trunk?

	PR c++/85043

gcc/cp/ChangeLog:

	* typeck.cc (maybe_warn_about_useless_cast): Don't warn when a class
	object is cast to a non-reference type.

gcc/ChangeLog:

	* doc/invoke.texi: Update documentation of -Wuseless-cast.

gcc/testsuite/ChangeLog:

	* g++.dg/warn/Wuseless-cast.C: Remove dg-warning.
	* g++.dg/warn/Wuseless-cast3.C: New test.