libstdc++: Refactor implementation of operator+ for std::string
Sorry for the delay. Tested on x86-64 Linux.

-->8--

After consultation with Jonathan, it seemed like a good idea to create a
single function that performed one-allocation string concatenation that
could be used by various different version of operator+. This patch adds
such a function and calls it from the relevant implementations.

libstdc++-v3/ChangeLog:

	* include/bits/basic_string.h:
	Add common function that performs single-allocation string
	concatenation. (__str_cat)
	Use __str_cat to perform optimized operator+, where relevant.
	* include/bits/basic_string.tcc::
	Remove single-allocation implementation of operator+.

Signed-off-by: Will Hawkins <whh8b@obs.cr>