Always use TYPE_MODE instead of DECL_MODE for vector field
commit e034c5c895722e0092d2239cd8c2991db77d6d39
Author: Jakub Jelinek <jakub@redhat.com>
Date:   Sat Dec 2 08:54:47 2017 +0100

	PR target/78643
	PR target/80583
	* expr.c (get_inner_reference): If DECL_MODE of a non-bitfield
	is BLKmode for vector field with vector raw mode, use TYPE_MODE
	instead of DECL_MODE.

fixed the case where DECL_MODE of a vector field is BLKmode and its
TYPE_MODE is a vector mode because of target attribute.  Remove the
BLKmode check for the case where DECL_MODE of a vector field is a vector
mode and its TYPE_MODE is BLKmode because of target attribute.

gcc/

	PR target/107304
	* expr.c (get_inner_reference): Always use TYPE_MODE for vector
	field with vector raw mode.

gcc/testsuite/

	PR target/107304
	* gcc.target/i386/pr107304.c: New test.